+++ src/main/java/umc/study/domain/Review.java
@@ -2,7 +2,7 @@ package umc.study.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.study.domain.commom.BaseEntity;
+import umc.study.domain.common.BaseEntity;
 
 @Entity
 @Getter
